Removed unnecessary code snippets

+++ src/app/core/services/products/product.service.ts
@@ -1,14 +1,13 @@
 import { Injectable, signal } from '@angular/core';
 import { Products } from '../../interfaces/products.interface';
-import { Observable, of } from 'rxjs';
 import { InventoryService } from '../inventory/inventory.service';
+import { getMatIconFailedToSanitizeLiteralError } from '@angular/material/icon';
 
 @Injectable({
   providedIn: 'root'
 })
 export class ProductService {
   
-  private lastId = signal(this.getLastId() || 0);
   public isEditting = signal<boolean>(false);
 
   constructor(private inventoryService: InventoryService) { }
@@ -23,40 +22,43 @@ export class ProductService {
   }
 
   public generateId(): number{
-    this.lastId.update(id => id + 1);
-    const newProductId = this.lastId();
+    let newProductId = this.getLastId() + 1;
     this.setLastProductId(newProductId);
     return newProductId;
   }
 
-  public getAllProducts():Observable<Products[]>{
+  public getAllProducts():Products[]{
     let products = JSON.parse(localStorage.getItem('allProducts') || '[]') as Products[];
-    return of(products);
+    return products;
   }
 
-  public createProduct(newProduct: Products):Observable<Products>{
-    let allProducts = JSON.parse(localStorage.getItem('allProducts') || '[]') as Products[];
-    allProducts.push(newProduct);
-    localStorage.setItem('allProducts',JSON.stringify(allProducts));
-    return of(newProduct);
+  public createProduct(newProduct: Products):boolean{
+    let allProducts = this.getAllProducts();
+    let productExists = allProducts.find(product => product.name === newProduct.name);
+    if(productExists){
+      return false;
+    } else {
+      allProducts.push(newProduct);
+      localStorage.setItem('allProducts',JSON.stringify(allProducts));
+      return true;
+    }
   }
 
-  public updateProduct(updatedProductData: Products):Observable<boolean>{
+  public updateProduct(updatedProductData: Products):boolean{
     let allProducts = JSON.parse(localStorage.getItem('allProducts') || '[]') as Products[];
     let productIndex = allProducts.findIndex(product => product.id === updatedProductData.id);
     if(productIndex !== -1){
       allProducts[productIndex] = updatedProductData;
       localStorage.setItem('allProducts', JSON.stringify(allProducts));
-      return of(true);
+      return true;
     }
-    return of(false);
+    return false;
   }
 
-  public deleteProduct(categoryId:number, productId: number):Observable<boolean>{
+  public deleteProduct(categoryId:number, productId: number){
     let allProducts = JSON.parse(localStorage.getItem('allProducts') || '[]') as Products[];
     allProducts = allProducts.filter(product => product.id !== productId);
     localStorage.setItem('allProducts',JSON.stringify(allProducts));
     this.inventoryService.deleteProductFromCategory(categoryId, productId);
-    return of(true);
   }
 }
